requirements.txt
beautifulsoup4==4.11.2
Flask==2.2.3
Flask-SocketIO
lxml==4.9.2
python-engineio==3.14.2
python-socketio==4.6.0
requests==2.28.2
uwsgi; sys_platform == "linux"